% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_correction.R
\name{add_gc_cor}
\alias{add_gc_cor}
\title{Add GC correction to SCE Object}
\usage{
add_gc_cor(
  sce,
  gc = rowData(sce)$gc,
  assay_name = "counts",
  method = c("modal", "copykit", "loess"),
  verbose = FALSE,
  ncores = 1,
  ...
)
}
\arguments{
\item{sce}{an SCE object}

\item{gc}{GC corresponding to bins (rows) in the matrix}

\item{assay_name}{Name of assay on which to perform GC correction}

\item{method}{Specifies the type of GC correction to perform. One of \verb{'modal', 'copykit', or 'loess'}}

\item{verbose}{Message verbosity (TRUE/FALSE)}

\item{ncores}{Number of cores to use if parallel backend is available}

\item{...}{Additional arguments to be passed to GC correction methods}
}
\value{
sce object with corrected GC count matrix in \code{assay(sce, 'counts_gc_[method]')}. See
}
\description{
Add GC correction to SCE Object
}
