% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{add_gc_freq}
\alias{add_gc_freq}
\title{Add GC frequency}
\usage{
add_gc_freq(bs_genome, bins)
}
\arguments{
\item{bs_genome}{BSGenome object}

\item{bins}{GRanges bins object}
}
\value{
GRanges bin object with GC and N frequency per bin
}
\description{
Add GC frequency
}
