% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmcopy_utils.R
\name{add_hmmcopy}
\alias{add_hmmcopy}
\title{Add HMMCopy results to sce object}
\usage{
add_hmmcopy(
  sce,
  verbose = FALSE,
  ncores = 1,
  assay_name = paste0("counts_gc_", sce@metadata$gc_cor_method),
  save_raw_hmm = NULL,
  slot_suffix = NULL,
  ...
)
}
\arguments{
\item{sce}{an SCE object}

\item{verbose}{Message verbosity}

\item{ncores}{Number of cores}

\item{assay_name}{Name of the assay with counts to input into HMMcopy. Ideally these are GC corrected.}

\item{save_raw_hmm}{Path to save raw hmm data in an \code{rda} file}

\item{slot_suffix}{Suffix to add to newly created \code{copy} and \code{state} assay slots.}

\item{...}{Additional parameters to pass to \link{run_sc_hmmcopy}, such as \code{param}}
}
\value{
an sce object with hmm copy metadata added to coldata, and new slots \code{copy} and \code{state}
}
\description{
Add HMMCopy results to sce object
}
\examples{
data(test_sce)
test_sce <- add_hmmcopy(test_sce)
}
