% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{bin_frags}
\alias{bin_frags}
\title{Bin fragments from ArrowFile}
\usage{
bin_frags(
  ArrowFile,
  bins,
  outfile = NULL,
  ncores = 1,
  BPPARAM = BiocParallel::MulticoreParam(workers = ncores, progressbar = TRUE),
  ...
)
}
\arguments{
\item{ArrowFile}{ArrowFile}

\item{bins}{Bins GRanges object}

\item{outfile}{Optional: Path to write output \code{.mtx} file}

\item{ncores}{Number of cores to use}

\item{BPPARAM}{Options to \code{BPPARAM} to enable multithreading}

\item{...}{Additional arguments passed to \code{bplapply}}
}
\value{
Binned depth sparse matrix
}
\description{
Parallel enabled depth counting of read fragments in given genomic bins
}
