% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{bin_frags_chr}
\alias{bin_frags_chr}
\title{Bin scATAC fragments}
\usage{
bin_frags_chr(fragments_chr, bins, blacklist = NULL)
}
\arguments{
\item{fragments_chr}{A single chromosome fragments}

\item{bins}{A GRanges bins (can include all chromosomes)}

\item{blacklist}{Genome blacklist regions to filter against}
}
\value{
Sparse matrix of binned fragment counts
}
\description{
\code{bin_frags_chr} computes the fragments across bins in a single chromosome from an ArchR ArrowFile
}
