% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_snp_data.R
\name{bin_snp_data}
\alias{bin_snp_data}
\title{Bin SNPs into genomic segments}
\usage{
bin_snp_data(
  snp,
  tilewidth = 10000000,
  respect_chr_arms = FALSE,
  select_chrs = NULL,
  bs_genome = NULL,
  bins = get_tiled_bins(bs_genome = bs_genome, tilewidth = tilewidth, select_chrs =
    select_chrs, respect_chr_arms = respect_chr_arms)
)
}
\arguments{
\item{snp}{SingleCellExperiment SNP object generated by \link{read_vartrix}}

\item{tilewidth}{Bin size}

\item{respect_chr_arms}{logical If \code{TRUE}, bins will be created with respect to chromosome arms (ie. not crossing arm boundries)}

\item{select_chrs}{Vector of chromosomes to include}

\item{bs_genome}{A BSgenome object}

\item{bins}{GRanges object with genome bins. Automatically generated by \link{get_tiled_bins} if not provided.}
}
\value{
SCE object with binned SNPs
}
\description{
Bin SNPs into genomic segments
}
