% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_atac_bins.R
\name{bin_snp_data2}
\alias{bin_snp_data2}
\title{Bin SNP Data}
\usage{
bin_snp_data2(snp_sce, binsize = 500000, select_chrs = NULL, bins = NULL)
}
\arguments{
\item{snp_sce}{SCE with snp data}

\item{binsize}{Size of bins}

\item{select_chrs}{Chromosomes to include}

\item{bins}{Optional override of bins}
}
\value{
SCE object with phased binned snps
}
\description{
Bin SNP Data
}
