% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_sublist}
\alias{bind_sublist}
\title{Bind sublists within lists}
\usage{
bind_sublist(
  toplist,
  sublist,
  what = c("rbind"),
  .add_id = FALSE,
  .id_name = "id"
)
}
\arguments{
\item{toplist}{List of lists of \code{rbind}able data}

\item{sublist}{Name or index of sublist}

\item{what}{Specify either \code{cbind} or \code{rbind} (Currently only rbind implemented)}

\item{.add_id}{Add identifiers to each data entry prior to merging. Based on list names if available. (default=FALSE)}

\item{.id_name}{Name of the id column if added}
}
\description{
Helper function for lists of lists, where the top level list is over multiple samples, and the sublists are results that are identical across samples that one wishes to bind
}
\examples{
toplist <- list(
  sample_1 = list(
    result1 = data.frame(
      a = c(1, 2, 3),
      b = c("X", "Y", "Z")
    ),
    result2 = data.frame(
      height = 180,
      weight = 75
    )
  ),
  sample_2 = list(
    result1 = data.frame(
      a = c(6, 5, 4),
      b = c("A", "B", "C")
    ),
    result2 = data.frame(
      height = 155,
      weight = 60
    )
  )
)

bind_sublist(toplist, sublist = 1, what = "rbind", .add_id = T)

bind_sublist(toplist, sublist = "result2", what = "rbind", .add_id = T)

bind_sublist(toplist, sublist = 2, what = "rbind", .add_id = F)

}
