% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bins_10mb}
\alias{bins_10mb}
\title{10mb Bins for HG38}
\format{
\subsection{\code{bins_10mb}}{

A GenomicRanges object with 331 bins.
}
}
\usage{
bins_10mb
}
\description{
A subset of cells from normal mammary
}
\keyword{datasets}
