% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allelic_imbalance.R
\name{calc_ai}
\alias{calc_ai}
\title{Calculate allelic imbalance}
\usage{
calc_ai(ref_counts, alt_counts)
}
\arguments{
\item{ref_counts}{vector of ref counts}

\item{alt_counts}{vector of alt counts}
}
\value{
measure of allelic imbalance
}
\description{
Calculate allelic imbalance
}
\examples{
calc_ai(c(4, 2, 1), c(2, 3, 1))
}
