% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allelic_imbalance.R
\name{calc_allelic}
\alias{calc_allelic}
\title{Calculate allelic information over cells and bins}
\usage{
calc_allelic(snp, ncores = 1, bins = NULL, group_var = NULL, FUN = calc_ai)
}
\arguments{
\item{snp}{SNP SingleCellExperiment object with slots 'ref' and 'alt' per cell counts}

\item{ncores}{Number of cores to speed up computation}

\item{bins}{Bins over which to aggregate allele information. If not provided will calculate for each SNP}

\item{group_var}{Column containing cell grouping information}

\item{FUN}{A function which can take two values (or vectors) of reference and alternate counts to produce another value (or vector)}
}
\value{
a snp sce
}
\description{
Calculate allelic information over cells and bins
}
