% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clonal_diversity.R
\name{calc_clonal_diversity}
\alias{calc_clonal_diversity}
\title{Calculate clonal diversity}
\usage{
calc_clonal_diversity(sce, sample_column, clone_name, index = "shannon")
}
\arguments{
\item{sce}{sce}

\item{sample_column}{sample}

\item{clone_name}{clone}

\item{index}{diversity index to use. One of \code{"shannon"}, \code{"simpson"}, or
\code{"invsimpson"}.}
}
\value{
named list
}
\description{
Calculate clonal diversity
}
