% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_cnv_score}
\alias{calc_cnv_score}
\title{Calculate CNV Score}
\usage{
calc_cnv_score(sce, assay_name = "counts")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay_name}{Name of assay}
}
\value{
SCE object with \code{cnv_score} appended
}
\description{
Calculate CNV Score
}
