% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_snn_specificity.R
\name{calc_snn_specificity}
\alias{calc_snn_specificity}
\title{Calculate SNN Specificity}
\usage{
calc_snn_specificity(
  snn_graph,
  cell_idents1,
  cell_idents2,
  sampling_rate = 0.1,
  ncores = 1
)
}
\arguments{
\item{snn_graph}{A SNN matrix}

\item{cell_idents1}{Vector of cell identities for which the specificity will be calculated against. Ideally sample or patient.}

\item{cell_idents2}{Vector of cell identities for which define cell types or states. The variable for which we want to see if the factor is specific to cell_idents1.}

\item{sampling_rate}{Proportion of cells to subsample when performing calculations.}

\item{ncores}{Number of cores to use}
}
\value{
A dataframe
}
\description{
Calculates the fraction of nearest neighbours divided by the expected fraction of nearest neighbours in the patient subgraph.
}
