% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allelic_imbalance.R
\name{calc_snp_ai}
\alias{calc_snp_ai}
\title{Calculate SNP allelic imbalance}
\usage{
calc_snp_ai(snp, ncores = 1)
}
\arguments{
\item{snp}{SNP SingleCellExperiment object}

\item{ncores}{Number of cores to speed up computation}
}
\value{
a snp sce
}
\description{
Calculate SNP allelic imbalance
}
