% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{clone_cna_comp_plot}
\alias{clone_cna_comp_plot}
\title{Clone Copy Number Comparison Plot}
\usage{
clone_cna_comp_plot(
  sce,
  clone_column = "clusters",
  subset_clones = NULL,
  subset_chr = NULL,
  assay_name = "segment_merged_logratios",
  center_point = 0,
  pseudobulk_fun = mean
)
}
\arguments{
\item{sce}{sce object}

\item{clone_column}{column name with clone information}

\item{subset_clones}{optional vector of clones to subset for comparison}

\item{subset_chr}{optional vector of chromosomes to subset for comparison}

\item{assay_name}{name of assay to use}

\item{center_point}{Center point of the assay for drawing dashed lines}

\item{pseudobulk_fun}{mean or median}
}
\value{
ggapply plot
}
\description{
Clone Copy Number Comparison Plot
}
