% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cnaHeatmap}
\alias{cnaHeatmap}
\alias{cloneCnaHeatmap}
\title{Plot copy number heatmap}
\usage{
cnaHeatmap(
  sce,
  assay_name = "state",
  clone_name = NULL,
  cluster_cells = FALSE,
  log2 = FALSE,
  center = FALSE,
  scale = c("none", "cells", "bins", "both"),
  label_genes = NULL,
  col_fun = NULL,
  col_clones = NULL,
  cluster_clones = FALSE,
  legend_name = assay_name,
  clust_annot = TRUE,
  verbose = TRUE,
  top_annotation = NULL,
  bulk_cn_col = NULL,
  raster_quality = 10,
  use_raster = TRUE,
  row_split = NULL,
  ...
)

cloneCnaHeatmap(
  sce,
  assay_name = "counts",
  clone_name = NULL,
  scale = c("none", "cells", "bins", "both"),
  log2 = FALSE,
  center = FALSE,
  clust_lab = TRUE,
  aggr_fun = mean,
  round = FALSE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{assay_name}{Name of the assay to plot}

\item{clone_name}{Name of clone_id column in sce object}

\item{cluster_cells}{If the value is a logical, it controls whether to make cluster on rows. The value can also be a \code{stats::hclust} or a \code{stats::dendrogram} which already contains clustering. Note this will override ordering of specified clones. Check https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#clustering.}

\item{log2}{Logical: Log2 transform the matrix prior to plotting}

\item{center}{logical: center the matrix prior to plotting}

\item{scale}{One of \verb{'cells', 'bins', 'both' or 'none'}. Determines what kind of scaling is done.}

\item{label_genes}{Optional: Vector of gene names to label in the heatmap. Note: \link{overlap_genes} must be run prior to labelling genes.}

\item{col_fun}{Color mapping function from \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}}

\item{col_clones}{Optional: A named vector (or unnamed) of clone colors.}

\item{cluster_clones}{Logical: Whether or not to order the clones by clustering}

\item{legend_name}{Name of the legend}

\item{clust_annot}{Annotate cluster and sample labels}

\item{verbose}{Logical: Message verbosity}

\item{top_annotation}{Top annotation as per \code{\link[ComplexHeatmap:columnAnnotation]{ComplexHeatmap::columnAnnotation()}}}

\item{bulk_cn_col}{Name of column in \code{rowRanges(sce)} that contains bulk copy number data to plot on top of heatmap}

\item{raster_quality}{Quality of raster (default: 10)}

\item{use_raster}{Logical: rasterize the heatmap body via \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}} (default: \code{TRUE}). Set to \code{FALSE} on systems without a working raster device (e.g. some headless CI runners).}

\item{row_split}{Row split for \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}}

\item{...}{Additional parameters that can be passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}}

\item{clust_lab}{Cluster labels}

\item{aggr_fun}{Aggregation function}

\item{round}{Round pseudobulked values to integers}
}
\value{
A heatmap
}
\description{
Plot copy number heatmap
}
