% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{filter_sce}
\alias{filter_sce}
\title{Bin and/or Cell-wise quality filtering}
\usage{
filter_sce(
  sce,
  assay_name = "counts",
  which = c("bins", "cells"),
  min_counts = 1,
  min_prop = 0.95,
  flag_only = FALSE,
  gc_range = c(-Inf, Inf)
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay_name}{Name of assay to pull counts from. Ideally raw counts}

\item{min_counts}{Minimum counts per cell/bin}

\item{min_prop}{Minimum proportion of cells/bins with \code{min_counts}}

\item{flag_only}{Only flag cells/bins, do not remove}

\item{gc_range}{Min and max GC values per bin to keep}

\item{filt}{Filter cells or bins}
}
\value{
SingleCellExperiment object
}
\description{
Note -- this filters GC bins, then bins by counts, and then cells. So the denominator changes at each step.
}
