% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{flagDoublets}
\alias{flagDoublets}
\title{Flag doublets}
\usage{
flagDoublets(
  sce,
  cutEnrich = 1,
  cutScore = -Inf,
  filterRatio = 1,
  remove = FALSE
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{cutEnrich}{The minimum numeric cutoff for \code{DoubletEnrichment}. This number is equivalent to the number of simulated
doublets identified as a nearest neighbor to the cell divided by the expected number given a random uniform distribution.}

\item{cutScore}{The minimum numeric cutoff for \code{DoubletScore} which represents the \verb{-log10(binomial adjusted p-value)} for the \code{DoubletEnrichment}.}

\item{filterRatio}{The maximum ratio of predicted doublets to filter based on the number of pass-filter cells.
For example, if there are 5000 cells, the maximum would be \code{filterRatio * 5000^2 / (100000)} (which simplifies to \code{filterRatio * 5000 * 0.05}).
This \code{filterRatio} allows you to apply a consistent filter across multiple different samples that may have different
percentages of doublets because they were run with different cell loading concentrations.
The higher the \code{filterRatio}, the greater the number of cells potentially removed as doublets.}

\item{remove}{Logical. Whether to remove doublets from the object.}
}
\description{
This function is modified from \code{ArchR::filterDoublets} to enable flagging of doublets prior to removal.
}
