% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_correction.R
\name{gc_cor_modal}
\alias{gc_cor_modal}
\title{Modal regression GC Correction for single cell data}
\usage{
gc_cor_modal(
  counts,
  gc,
  valid = rep(TRUE, length(counts)),
  bin_ids = names(counts),
  lowess_frac = 0.2,
  q = c(0.1, 0.9),
  g = c(0.1, 0.9),
  results = c("counts", "default", "full")
)
}
\arguments{
\item{counts}{Vector of counts in bins}

\item{gc}{Vector of GC content in bins}

\item{valid}{Vector of values corresponding to valid bins. If none provided defaults to all TRUE}

\item{bin_ids}{Bin ids}

\item{lowess_frac}{The smoother span. See \code{\link[stats:lowess]{stats::lowess()}} for more details}

\item{q}{A tuple of quantile bounds to compute quantile regressions over read counts}

\item{g}{GC quantiles over which to integrate regression curves}

\item{results}{Format of results to return. One of \code{full}, \code{default}, or \code{counts}.}
}
\value{
A dataframe with the modal quantile and GC corrected counts. Results are returned based on the \code{results} parameter as follows:
\describe{
\item{full}{Full dataframe containing results from each quantile curve}
\item{default}{Condensed dataframe from the selected modal quantile curve}
\item{counts}{Return only a vector of corrected counts}
}
}
\description{
Modal regression GC Correction for single cell data
}
