% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_assay_dat}
\alias{get_assay_dat}
\title{Utility to pull assay data and merge with features}
\usage{
get_assay_dat(sce, assay_names, cell_id = colnames(sce))
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{cell_id}{Cell id to grab data}

\item{assays}{Vector list of assays}
}
\value{
data frame with merged data
}
\description{
Utility to pull assay data and merge with features
}
