% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blacklist.R
\name{get_blacklist}
\alias{get_blacklist}
\title{Get blacklist regions}
\usage{
get_blacklist(genome = NULL)
}
\arguments{
\item{genome}{Genome version: hg38, hg19, mm10, ce10, or dm3}
}
\value{
Genomic ranges of blacklist regions
}
\description{
Get blacklist regions
}
