% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{get_cancer_gene_copy}
\alias{get_cancer_gene_copy}
\title{Get Cancer Gene Copies}
\usage{
get_cancer_gene_copy(sce, assay_name, group_var = "all")
}
\arguments{
\item{sce}{Single cell experiment object}

\item{assay_name}{Name of the assay from which to pull copy number}

\item{group_var}{Grouping variable (defaults to all)}
}
\value{
Data frame containing gene copies
}
\description{
Calculate per group copy number for a list of cancer genes.
}
