% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{get_chr_arm_bins}
\alias{get_chr_arm_bins}
\title{Get chromosome arm bins}
\usage{
get_chr_arm_bins(genome = "hg38")
}
\arguments{
\item{genome}{Genome version ('hg38', 'hg19')}
}
\value{
A GRanges object of chromosome arm bins
}
\description{
Get chromosome arm bins
}
\examples{
bins <- get_chr_arm_bins('hg38')
}
