% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{get_ideal_mat}
\alias{get_ideal_mat}
\title{Get ideal bin matrix}
\usage{
get_ideal_mat(
  mat,
  gc,
  n_freq,
  map,
  min_reads = 1,
  max_N_freq = 0.05,
  reads_outlier = 0.01,
  gc_outlier = 0.001,
  min_map = 0.9,
  ncores = 1
)
}
\arguments{
\item{mat}{Count matrix}

\item{gc}{Vector of gc content}

\item{n_freq}{Vector of bin N frequency (proportion of N bases in a bin)}

\item{map}{Vector of bin mappability}

\item{min_reads}{Minimum number of reads to consider a bin}

\item{max_N_freq}{Maximum allowable frequency of N bases to consider a bin. Range (0, 1)}

\item{reads_outlier}{Flag bins with reads in the top quantile given by this value. Range (0, 1)}

\item{gc_outlier}{Flag bins with GC content in the top and bottom quantule given by this value. Range (0, 1)}

\item{min_map}{Minimum allowable mappability score for a bin. Range (0, 1)}

\item{ncores}{number of cores for parallel evaluation (requires \link{parallel} package)}
}
\value{
A logical vector of same length as input identifying which bins in a given cell meet the filtering criteria
}
\description{
Given a matrix of bin counts, bin gc and N frequency, and filtering parameters, return a boolean matrix flagging ideal bins
}
