% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_snp_counts}
\alias{get_snp_counts}
\title{get_snp_counts}
\usage{
get_snp_counts(sce, variables = "all", target_assays = c("ref", "alt"))
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object with one or more assays}

\item{variables}{Factors upon which to split the counts. Defaults to \code{'all'} which counts across entire dataset}

\item{target_assays}{Assays in \code{sce} to count}
}
\value{
A \code{SingleCellExperiment} object with named counts for each assay in the \code{rowData(sce)} slot.
}
\description{
Count ref and alt snps in sce.
}
\examples{
\dontrun{
sce <- get_snp_counts(sce, variables = c("all", "Sample", "Condition"), target_assays = c("ref", "alt"))
}
}
