% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{get_tiled_bins}
\alias{get_tiled_bins}
\title{Get tiled bins}
\usage{
get_tiled_bins(
  bs_genome = NULL,
  tilewidth = 500000,
  select_chrs = NULL,
  respect_chr_arms = TRUE
)
}
\arguments{
\item{bs_genome}{A BSgenome object}

\item{tilewidth}{Bin size}

\item{select_chrs}{Vector of chromosomes to include}

\item{respect_chr_arms}{logical If \code{TRUE}, bins will be created with respect to chromosome arms (ie. not crossing arm boundries)}
}
\value{
A GRanges object of bins
}
\description{
Get tiled bins
}
\examples{
\dontrun{
bins <- get_tiled_bins(BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38, tilewidth = 500000)
}
}
