% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmcopy_utils.R
\name{grab_hmm_res}
\alias{grab_hmm_res}
\title{Grab HMM results from returned list}
\usage{
grab_hmm_res(hmm_results, grab = "best", ncores = 1)
}
\arguments{
\item{hmm_results}{hmm results in multilist format}

\item{grab}{What to grab ("best" for best results; m1, m2, m3...for specific multiplier)}

\item{ncores}{number of cores for parallelization}
}
\value{
A list of each cells best result\
}
\description{
Helper function to parse large list of HMM results
}
\details{
Note: Parallel currently disabled as it seems slower than single core
}
