% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmcopy_utils.R
\name{hmmcopy_singlecell}
\alias{hmmcopy_singlecell}
\title{Single Cell HMMcopy}
\usage{
hmmcopy_singlecell(
  chr,
  start,
  end,
  counts,
  reads,
  param = sc_hmm_params(),
  cell_id,
  multiplier = 1,
  verbose = FALSE,
  maxiter = 200,
  n_cutoff = NULL
)
}
\arguments{
\item{chr}{Vector of chromosomes}

\item{start}{Vector of bin start positions}

\item{end}{Vector of bin end positions}

\item{counts}{Vector of bin corrected counts. These should ideally be GC corrected counts from \code{\link[=modal_quantile_regression]{modal_quantile_regression()}}}

\item{reads}{A vector of raw read counts per bin}

\item{param}{A matrix of parameter values in columns for each state in rows. See \code{\link[HMMcopy:HMMsegment]{HMMcopy::HMMsegment()}} for more information.}

\item{cell_id}{Cell id}

\item{multiplier}{Ploidy multiplier}

\item{verbose}{Print verbose}

\item{maxiter}{The maximum number of iterations allows for the Maximum-Expectation algorithm, reduce to decrease running time at the expense of robustness.}

\item{n_cutoff}{Cutoff for the number of bins in a given state for calculating the \code{true_multiplier} value. Defaults to 5\% of bins.}
}
\value{
A list with the following objects:
\describe{
\item{bincounts}{Data frame of bin counts with copy state}
\item{modal_seg}{Data frame of genome segments}
\item{mstats}{A single row of cell summary statistics}
\item{df_params}{Data frame of parameters used for each iteration and state}
}
}
\description{
Runs \code{HMMCopy} on single cell binned counts.
}
