% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_tools.R
\name{identify_normal}
\alias{identify_normal}
\title{Identify normal cells from scATAC data}
\usage{
identify_normal(
  sce,
  assay_name,
  group_by = "clusters",
  method = c("gmm", "min_sd"),
  n_normal_clusts = NULL,
  plot = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment Object}

\item{assay_name}{Name of assay to calculate variance from}

\item{group_by}{Name of column containing the grouping information}

\item{method}{One of \code{gmm} or \code{min_sd}}

\item{n_normal_clusts}{Number of expected normal clusters (only for method \code{min_sd})}
}
\value{
SingleCellExperiment object with column \code{tumor_cell}
}
\description{
Uses the standard deviation of the difference between each bin to estimate tumor and normal cell clusters. Using method \code{gmm} will calculate the per cluster median of the sd, and then fit a two-component GMM to determine tumor cells. If method \code{min_sd} is specified (or if only two clusters are provided), simply uses the \code{n_normal_clusts} with the lowest median sd.
}
