% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_segments.R
\name{integrate_segments}
\alias{integrate_segments}
\title{Integrate segments between assays}
\usage{
integrate_segments(x, y, granges_signal_colname, drop_na = TRUE)
}
\arguments{
\item{x}{granges object to integrate on}

\item{y}{granges}

\item{granges_signal_colname}{column containing cna signal in granges}

\item{drop_na}{Logical. Drop bins with any NA values.}
}
\value{
\code{x} with averaged signal of \code{y} integrated over the bins
}
\description{
Will merge bins of \code{y} into \code{x}, taking the weighted mean of the binned signal in \code{y} of \code{granges_signal_colname}
}
