% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{is_ideal_bin}
\alias{is_ideal_bin}
\title{Flag ideal bins}
\usage{
is_ideal_bin(
  counts,
  gc,
  n_freq,
  map = NULL,
  min_reads = 0,
  max_N_freq = 0.05,
  reads_outlier = 0.01,
  gc_outlier = 0.001,
  min_map = 0.9
)
}
\arguments{
\item{counts}{Vector of bin counts for single cell}

\item{gc}{Vector of gc content}

\item{n_freq}{Vector of bin N frequency (proportion of N bases in a bin)}

\item{map}{Vector of bin mappability}

\item{min_reads}{Minimum number of reads to consider a bin}

\item{max_N_freq}{Maximum allowable frequency of N bases to consider a bin. Range (0, 1)}

\item{reads_outlier}{Flag bins with reads in the top quantile given by this value. Range (0, 1)}

\item{gc_outlier}{Flag bins with GC content in the top and bottom quantule given by this value. Range (0, 1)}

\item{min_map}{Minimum allowable mappability score for a bin. Range (0, 1)}
}
\value{
A dataframe of two columns meet the \code{valid} or \code{ideal} criteria
}
\description{
\code{is_ideal_bin} will apply a set of bin-wise filters, based on high count outliers, high or low gc outliers, minimum read counts, minimum mappability, or maximum allowable frequency of N bases per bin.
}
