% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{length_normalize}
\alias{length_normalize}
\title{Length normalize counts}
\usage{
length_normalize(
  sce,
  assay_name = "counts",
  assay_to = "counts_lenNorm",
  binwidths = width(rowRanges(sce)),
  by_factor = getmode(binwidths),
  verbose = FALSE
)
}
\arguments{
\item{sce}{SCE object}

\item{assay_name}{Name of assay to normalize}

\item{by_factor}{Multiplication factor for counts}

\item{verbose}{Print verbose (TRUE/FALSE)}

\item{binwidth}{Vector of binwidths}
}
\value{
An sce object with counts length normalized in \code{assay(sce, 'counts_permb')}
}
\description{
By default this procedure will only impact counts in the bins that are variable length (for example tail ends of chromosomes).
}
