% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_atac_bins.R
\name{load_atac_bins}
\alias{load_atac_bins}
\title{Load atac binned depth data}
\usage{
load_atac_bins(
  samples,
  sample.names,
  ArchR_Proj = NULL,
  bins = NULL,
  BPPARAM = BiocParallel::bpparam(),
  save_to = NULL,
  verbose = FALSE,
  save_as = c("sce", "adata", "seurat")
)
}
\arguments{
\item{samples}{A character vector containing one or more directory names, each corresponding to a 10X sample.
Each directory should contain a matrix file, a gene/feature annotation file, and a barcode annotation file.

Alternatively, each string may contain a path to a HDF5 file in the sparse matrix format generated by 10X.
These can be mixed with directory names when \code{type="auto"}.

Alternatively, each string may contain a prefix of names for the three-file system described above,
where the rest of the name of each file follows the standard 10X output.}

\item{sample.names}{A character vector of length equal to \code{samples}, containing the sample names to store in the column metadata of the output object.
If \code{NULL}, the file paths in \code{samples} are used directly.}

\item{ArchR_Proj}{Optional: ArchR project with matching cells and metadata}

\item{bins}{Optional: GRanges bins object}

\item{BPPARAM}{Options to pass to \code{bplapply} for data loading. Provides minor speedup if loading many samples}

\item{save_to}{File path in which to save the final output. Note: Will still return the sce object for downstream analysis.}

\item{verbose}{Message verbosity (TRUE/FALSE)}

\item{save_as}{NOT IMPLEMENTED YET. Select file formats to save the object. Can provide multiple values}

\item{directory}{Folder containing samples}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Loads binned atac reads, merges cell-wise and bin-wise metadata, and performs QC.
}
