% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{logr_col_fun}
\alias{logr_col_fun}
\title{Log Ratio Colors}
\usage{
logr_col_fun(breaks = c(-2, 0, 2), colors = c("blue", "white", "red"))
}
\arguments{
\item{breaks}{Value breaks}

\item{colors}{Colors}
}
\value{
Color function
}
\description{
Easy color mappings for log2 color scales in \code{ComplexHeatmap}
}
