% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_to_sce.R
\name{merge.sparse}
\alias{merge.sparse}
\title{Merge sparse matrices}
\usage{
\method{merge}{sparse}(l)
}
\arguments{
\item{l}{List of sparse matrices}
}
\value{
A merged matrix
}
\description{
Can merge when rows are non matching
}
