% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_tools.R
\name{merge_segments}
\alias{merge_segments}
\title{Merge segment levels}
\usage{
merge_segments(
  sce,
  smooth_assay,
  segment_assay,
  new_assay = "segment_merged",
  ncores = 1
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{smooth_assay}{name of assay with smoothed counts}

\item{segment_assay}{name of assay with segmented counts}

\item{new_assay}{Name of the new assay}

\item{ncores}{Number of cores to use}
}
\value{
A \code{SingleCellExperiment} object
}
\description{
Merge segment levels
}
