% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal_gc_correction.R
\name{modal_quantile_regression}
\alias{modal_quantile_regression}
\title{Modal regression GC Correction}
\usage{
modal_quantile_regression(
  counts,
  gc,
  bin_ids = seq_along(counts),
  lowess_frac = 0.2,
  q = c(0.1, 0.9),
  g = c(0.1, 0.9),
  full_return = FALSE
)
}
\arguments{
\item{counts}{Vector of counts in bins}

\item{gc}{Vector of GC content in bins}

\item{bin_ids}{Bin ids}

\item{lowess_frac}{The smoother span. See \code{\link[stats:lowess]{stats::lowess()}} for more details}

\item{q}{A tuple of quantile bounds to compute quantile regressions over read counts}

\item{g}{GC quantiles over which to integrate regression curves}

\item{full_return}{Return each individual quantile curve in the return df (default: FALSE)}
}
\value{
A dataframe with the modal quantile and GC corrected counts
}
\description{
Modal regression GC Correction
}
