% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{overlap_genes}
\alias{overlap_genes}
\title{Overlap genes with bins}
\usage{
overlap_genes(sce, ensDb, gene_biotype = "all")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{ensDb}{EnsemblDb object such as \link{EnsDb.Hsapiens.v86}}

\item{gene_biotype}{Gene biotypes to overlap. Defaults to "all". Use "protein_coding" to keep only protein coding genes}
}
\value{
SingleCellExperiment object with gene overlaps in \code{sce@metadata$gene_overlaps}
}
\description{
Given an annotation object, this function will overlap genes with bins and place the results in the metadata slot \code{gene_overlaps}. This function will also attempt to annotate cancer genes using OncoKB.
}
