% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_correction.R
\name{perform_gc_cor}
\alias{perform_gc_cor}
\title{Perform GC Correction}
\usage{
perform_gc_cor(mat, gc, method = c("modal", "copykit", "loess"), ...)
}
\arguments{
\item{mat}{Count matrix for GC correction}

\item{gc}{GC corresponding to bins (rows) in the matrix}

\item{method}{Specifies the type of GC correction to perform. One of \verb{'modal', 'copykit', or 'loess'}}

\item{...}{Additional arguments to be passed to GC correction methods or \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}}
}
\description{
Performs GC correction using \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}} over a matrix of cell counts
}
\details{
Note: If using \code{modal} must pass \code{results="counts"} as an argument.
}
