% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_snps.R
\name{phase_snps}
\alias{phase_snps}
\title{EM phasing of SNPs}
\usage{
phase_snps(
  snp,
  bins,
  max_iter = 50,
  sub_cells = NULL,
  phases = NULL,
  ncores = 1,
  min_snp_cov = 5,
  min_bin_snps = 5,
  tol = 0.001
)
}
\arguments{
\item{snp}{\code{SingleCellExperiment} object containing hetSNPs as generated by \link{read_vartrix}}

\item{bins}{\code{GenomicRanges} object containing bins (or segments) over which to perform phasing}

\item{max_iter}{Maximum number of iterations for the EM}

\item{sub_cells}{Subset cells when phasing}

\item{phases}{Named vector of SNP phases.}

\item{ncores}{Number of cores to parallelize over bins}

\item{min_snp_cov}{Minimum total coverage per SNP to be used in phasing.}

\item{min_bin_snps}{Minimum number of covered SNPs within a bin to be used.}

\item{tol}{maximum liklihood stopping tolerance}
}
\value{
\code{SingleCellExperiment} object containing EM phased SNPs. Per bin values are stored in \code{altExp(snp, 'em')}. Per SNP results are appended to \code{rowRanges(snp)}
}
\description{
EM phasing of SNPs
}
