% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cell_cna}
\alias{plot_cell_cna}
\title{Plot Cell Copy Number}
\usage{
plot_cell_cna(
  sce,
  cell_id = NULL,
  assay_name = "counts",
  col_fun = NULL,
  linewidth = 1
)
}
\arguments{
\item{sce}{sce object}

\item{cell_id}{Vector of cell ids}

\item{assay_name}{Name of assay to plot}

\item{col_fun}{Color mapping function from \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}}

\item{linewidth}{Width of the line segments in plot}
}
\value{
A ggplot object
}
\description{
Plot Cell Copy Number
}
