% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cell_multi}
\alias{plot_cell_multi}
\title{Plot multiple cell assays together}
\usage{
plot_cell_multi(sce, cell_id, assays)
}
\arguments{
\item{sce}{SCE object}

\item{cell_id}{Cell ids}

\item{assays}{Assays to plot}
}
\value{
List of ggplot objects
}
\description{
Plot multiple cell assays together
}
