% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clone_comp}
\alias{plot_clone_comp}
\title{Clone Copy Number Comparison Plot}
\usage{
plot_clone_comp(
  sce,
  group_var = "clusters",
  subset_clones = NULL,
  subset_chr = NULL,
  assay_name = "segment_merged_logratios",
  center_point = NULL,
  pt_size = 0.75,
  lg_pt_size = 3,
  pseudobulk_fun = mean
)
}
\arguments{
\item{sce}{sce object}

\item{group_var}{column name with clone information}

\item{subset_clones}{optional vector of clones to subset for comparison}

\item{subset_chr}{optional vector of chromosomes to subset for comparison}

\item{assay_name}{name of assay to use}

\item{center_point}{Center point of the assay for drawing dashed lines}

\item{pt_size}{Size of the points in the plot}

\item{lg_pt_size}{Size of the points in the legend}

\item{pseudobulk_fun}{mean or median}
}
\value{
ggapply plot
}
\description{
Clone Copy Number Comparison Plot
}
