% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_cna}
\alias{plot_gene_cna}
\title{Plot Gene CNA}
\usage{
plot_gene_cna(
  sce,
  gene,
  assay_type = "counts",
  group_by = NULL,
  color_by = group_by,
  return_data = FALSE
)
}
\arguments{
\item{sce}{sce}

\item{gene}{gene}

\item{assay_type}{assay}

\item{group_by}{grouping variable}

\item{color_by}{coloring variable}

\item{return_data}{If \code{TRUE}, will return plot data without plotting}
}
\value{
plot
}
\description{
Plot Gene CNA
}
