% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_segs}
\alias{plot_segs}
\title{Plot data with segments}
\usage{
plot_segs(sce, seg_assay, input_assay, cell_id)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object with data}

\item{seg_assay}{segmented assay}

\item{input_assay}{input data assay (pre-segmented)}

\item{cell_id}{Cell to plot}
}
\value{
ggplot
}
\description{
Will plot copy number tracks with raw input and overlaid segmented data
}
