% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{pseudo_groups}
\alias{pseudo_groups}
\title{Pseudobulk groups}
\usage{
pseudo_groups(sce, assay_name, group_var = NULL, FUN = mean, na.rm = TRUE, ...)
}
\arguments{
\item{sce}{sce obj}

\item{assay_name}{One or more assays to pseudobulk}

\item{group_var}{grouping variable}

\item{FUN}{funtion to use when pseudobulking}

\item{na.rm}{logical -- remove NAs in \code{FUN}}

\item{...}{additional parameters to pass to \code{FUN}}
}
\value{
A pseudobulked sce obj
}
\description{
Pseudobulk groups
}
