% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascn.R
\name{pseudo_join}
\alias{pseudo_join}
\title{Join SNP and scATAC depth}
\usage{
pseudo_join(
  sce,
  snp,
  sce_assay = "counts",
  ncores = 1,
  group_var = NULL,
  min_cov = 5,
  sce_aggr_fun = mean
)
}
\arguments{
\item{sce}{sce obj}

\item{snp}{snp obj}

\item{sce_assay}{sce_assay to aggr}

\item{ncores}{Number of cores to speed up computation}

\item{group_var}{Column containing cell grouping information}

\item{min_cov}{Minumum coverage per SNP across pseudobulked cells to be included in the calculation}

\item{sce_aggr_fun}{fx for aggr}
}
\value{
joint obj
}
\description{
Join SNP and scATAC depth
}
