% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scatools.R
\name{run_scatools}
\alias{run_scatools}
\title{Run scATAC Copy Number Analysis}
\usage{
run_scatools(
  sample_id,
  fragments = NULL,
  ArrowFile = NULL,
  outdir = sample_id,
  overwrite = FALSE,
  force_arrow = FALSE,
  verbose = TRUE,
  bins,
  ncores = 1
)
}
\arguments{
\item{sample_id}{Sample id}

\item{fragments}{Fragments file}

\item{ArrowFile}{ArrowFile}

\item{outdir}{Outdir}

\item{overwrite}{Whether to overwrite previous results}

\item{force_arrow}{recreate arrow file}

\item{verbose}{Verbosity}

\item{bins}{Bins granges}

\item{ncores}{Number of cores}
}
\value{
\code{SingleCellExperiment} object
}
\description{
Convenience wrapper to run a sample
}
