% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatools-package.R
\docType{package}
\name{scatools-package}
\alias{scatools}
\alias{scatools-package}
\title{scatools: scATAC tools and utilities}
\description{
A collection of useful code for handling scATAC data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjz1/scatools}
  \item \url{https://mjz1.github.io/scatools/}
  \item Report bugs at \url{https://github.com/mjz1/scatools/issues}
}

}
\author{
\strong{Maintainer}: Matthew Zatzman \email{matthew.zatzman@gmail.com}

}
\keyword{internal}
