% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_tools.R
\name{segment_cnv}
\alias{segment_cnv}
\title{scATAC CBS CNV segmentation}
\usage{
segment_cnv(
  sce,
  assay_name,
  new_assay = paste(assay_name, "segment", sep = "_"),
  alpha = 0.2,
  nperm = 10,
  min.width = 2,
  undo.splits = "none",
  arm_aware = TRUE,
  verbose = 0,
  bpparam = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{assay_name}{Name of the assay to segment}

\item{new_assay}{Name of the new assay}

\item{alpha}{significance levels for the test to accept change-points.}

\item{nperm}{number of permutations used for p-value computation.}

\item{min.width}{the minimum number of markers for a changed segment.
    The default is 2 but can be made larger.  Maximum possible value is
    set at 5 since arbitrary widths can have the undesirable effect of
    incorrect change-points when a true signal of narrow widths exists.}

\item{undo.splits}{A character string specifying how change-points are
    to be undone, if at all.  Default is "none".  Other choices are
    "prune", which uses a sum of squares criterion, and "sdundo", which 
    undoes splits that are not at least this many SDs apart.}

\item{arm_aware}{Logical. Segment within chromosome arms rather than whole
chromosomes when an \code{arm} annotation is present in \code{rowRanges(sce)}
(default \code{TRUE}).}

\item{verbose}{Verbosity}

\item{bpparam}{BiocParallel Parameters}

\item{...}{Additional parameters to pass to \link[DNAcopy:segment]{DNAcopy::segment}}
}
\value{
A \code{SingleCellExperiment} object
}
\description{
scATAC CBS CNV segmentation
}
