% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_tools.R
\name{smooth_counts}
\alias{smooth_counts}
\title{Smooth outlier copy number counts}
\usage{
smooth_counts(
  sce,
  assay_name,
  ncores = 1,
  smooth_name = paste(assay_name, "smoothed", sep = "_"),
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay_name}{Name of assay to smooth}

\item{ncores}{Number of cores to use}

\item{smooth_name}{Name of returned assay with smoothed counts}

\item{...}{}
}
\value{
A SingleCellExperiument Object
}
\description{
Smooth outlier copy number counts
}
