% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_chr_arm.R
\name{summarise_chr_arm}
\alias{summarise_chr_arm}
\title{Summarise copu number signal over chromosome arms}
\usage{
summarise_chr_arm(sce, assay_name, genome_ver = "hg38")
}
\arguments{
\item{sce}{SingleCellExperiment objet}

\item{assay_name}{Name of assay}

\item{genome_ver}{Genome version ("hg38" or "hg19")}
}
\value{
A chromosome arm level binned SingleCellExperiment
}
\description{
Summarise copu number signal over chromosome arms
}
