% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_sce}
\alias{test_sce}
\title{Processed scATAC dataset}
\format{
\subsection{\code{test_sce}}{

A SingleCellExperiment object with 255 bins and 67 cells.
}
}
\usage{
test_sce
}
\description{
A subset of cells from normal mammary
}
\keyword{datasets}
