% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmcopy_utils.R
\name{add_hmmcopy}
\alias{add_hmmcopy}
\title{Add HMMCopy results to sce object}
\usage{
add_hmmcopy(sce, verbose = FALSE, ncores = 1)
}
\arguments{
\item{sce}{an SCE object}

\item{verbose}{Message verbosity}

\item{ncores}{Number of cores}
}
\value{
an sce object with hmm copy metadata added to coldata, and new slots \code{copy} and \code{state}
}
\description{
Add HMMCopy results to sce object
}
\examples{
data(test_sce)
test_sce <- add_hmmcopy(test_sce)
}
