% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{bin_atac_frags}
\alias{bin_atac_frags}
\title{Bin atac fragments}
\usage{
bin_atac_frags(
  ArrowFiles,
  bins,
  blacklist = NULL,
  outdir,
  bin_name = prettyMb(getmode(width(bins))),
  ncores = 1,
  bpparams = BiocParallel::MulticoreParam(workers = ncores, progressbar = TRUE),
  overwrite = FALSE,
  return_mat = FALSE,
  ...
)
}
\arguments{
\item{ArrowFiles}{List or vector of ArrowFile paths}

\item{bins}{Bins GRanges object}

\item{outdir}{Optional: Directory to write the \code{.mtx}, \code{barcodes}, and \code{bins} files}

\item{bin_name}{Name of the bins (e.g. \code{'10Mb'}, \code{'500Kb'}, \code{'chr_arm'}). If not provided is automatically detected based on binwidth.}

\item{ncores}{Number of cores to use}

\item{bpparams}{Options to \code{BPPARAM} to enable multithreading}

\item{overwrite}{Logical. Overwrite previously existing results (default = FALSE)}

\item{return_mat}{Logical. Return the binned depth matrix (default = FALSE)}

\item{...}{Additional parameters passed to \code{bplapply}}
}
\value{
If \code{return_mat=TRUE}, returns a sparse binned depth matrix. Otherwise returns \code{NULL}
}
\description{
Convenience wrapper to bin fragments of a given size and save them as \code{.mtx} files
}
