% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{bin_frags}
\alias{bin_frags}
\title{Bin fragments from ArrowFile}
\usage{
bin_frags(
  fragment_file,
  bins,
  blacklist = NULL,
  outdir = NULL,
  ncores = 1,
  bpparams = BiocParallel::MulticoreParam(workers = ncores, progressbar = TRUE),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fragment_file}{Fragment file}

\item{bins}{Bins GRanges object}

\item{outdir}{Optional: Directory to write the \code{.mtx}, \code{barcodes}, and \code{bins} files}

\item{ncores}{Number of cores to use}

\item{bpparams}{Options to \code{BPPARAM} to enable multithreading}

\item{verbose}{Logical. Message verbosity}

\item{...}{Additional arguments passed to \code{bplapply}}
}
\value{
Binned depth sparse matrix
}
\description{
Parallel enabled depth counting of read fragments in given genomic bins from fragment files
}
