% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{bin_snp_data}
\alias{bin_snp_data}
\title{Title}
\usage{
bin_snp_data(snp_sce, binsize = 500000, select_chrs = NULL, bins = NULL)
}
\arguments{
\item{binsize}{Size of bins}

\item{select_chrs}{Chromosomes to include}

\item{bins}{Optional override of bins}

\item{snp_granges}{SNP granges object with GT column}
}
\value{
SCE object with phased binned snps
}
\description{
Title
}
