% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_snp_data.R, R/load_atac_bins.R
\name{bin_snp_data}
\alias{bin_snp_data}
\title{Bin SNPs into genomic segments}
\usage{
bin_snp_data(snp_sce, binsize = 5e+05, select_chrs = NULL, bins = NULL)

bin_snp_data(snp_sce, binsize = 5e+05, select_chrs = NULL, bins = NULL)
}
\arguments{
\item{binsize}{Size of bins}

\item{select_chrs}{Chromosomes to include}

\item{bins}{Optional override of bins}

\item{snp}{SingleCellExperiment SNP object generated by \link{read_vartrix}}

\item{snp_granges}{SNP granges object with GT column}
}
\value{
SCE object with binned SNPs

SCE object with phased binned snps
}
\description{
Bin SNPs into genomic segments

Title
}
