% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_allelic_imbalance.R
\name{calc_allelic}
\alias{calc_allelic}
\title{Calculate allelic information over cells and bins}
\usage{
calc_allelic(
  snp,
  ncores = 1,
  bins = NULL,
  group_var = NULL,
  FUN = calc_ai,
  min_cov = 1
)
}
\arguments{
\item{snp}{SNP SingleCellExperiment object with slots 'ref' and 'alt' per cell counts}

\item{ncores}{Number of cores to speed up computation}

\item{bins}{Bins over which to aggregate allele information. If not provided will calculate for each SNP}

\item{group_var}{Column containing cell grouping information}

\item{FUN}{A function which can take two values (or vectors) of reference and alternate counts to produce another value (or vector)}

\item{min_cov}{Minumum coverage per SNP across pseudobulked cells to be included in the calculation}
}
\value{
a snp sce with slots:
\itemize{
\item \code{res} - Containing the main result
\item \code{tot_cov} - Total SNP coverage in the bin for the set of cells after filtering for \code{min_cov}
\item \code{n_snp} - Number of SNPs remaining in the bin after filtering for \code{min_cov}
}
}
\description{
Calculate allelic information over cells and bins
}
