% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_cnv_score}
\alias{calc_cnv_score}
\title{Calculate CNV Score}
\usage{
calc_cnv_score(sce, assay_name = "counts", name = "cnv_score", method = "abs")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay_name}{Name of assay}

\item{name}{Name of new column to store cnv score}

\item{method}{'rms' for root mean square or 'abs' for absolute mean}
}
\value{
SCE object with \code{cnv_score} appended
}
\description{
Calculates the CNV score in a given assay using the absolute mean of CNV values.
}
