% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_seurat}
\alias{cluster_seurat}
\title{Cluster using Seurat}
\usage{
cluster_seurat(
  sce,
  assay_name,
  raw_counts = "counts",
  do.scale = FALSE,
  do.center = FALSE,
  algorithm = 1,
  resolution = 0.8,
  n.neighbors = 10,
  dims = 1:50,
  suffix = "",
  PCA_name = paste0("PCA", suffix),
  UMAP_name = paste0("UMAP", suffix),
  cluster_name = paste0("clusters", suffix),
  metric = "correlation",
  verbose = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay_name}{Assay name}

\item{raw_counts}{Assay containing raw counts}

\item{do.scale}{scale}

\item{do.center}{center}

\item{algorithm}{clustering algorithm}

\item{resolution}{clustering resolution}

\item{n.neighbors}{neighbors for umap}

\item{dims}{Number of reduced dimensions to use for FindNeighbors and UMAP}

\item{suffix}{Suffix name to add to the PCA, UMAP, and clusters}

\item{PCA_name}{Name to store PCA dimred}

\item{UMAP_name}{Name to store UMAP dimred}

\item{cluster_name}{Name to store seurat clusters}

\item{metric}{metric for umap}

\item{verbose}{Message verbosity}
}
\value{
SingleCellExperiment obj
}
\description{
Cluster using Seurat
}
