% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_seurat}
\alias{cluster_seurat}
\title{Cluster using Seurat}
\usage{
cluster_seurat(
  sce,
  assay_name,
  do.scale = FALSE,
  do.center = FALSE,
  algorithm = 1,
  resolution = 0.8,
  n.neighbors = 10,
  npcs.pca = 50,
  features.pca = "all",
  nvar.features = NULL,
  dims = 1:npcs.pca,
  k.param = 20,
  suffix = "",
  PCA_name = paste0("PCA", suffix),
  UMAP_name = paste0("UMAP", suffix),
  cluster_name = paste0("clusters", suffix),
  umap.metric = "correlation",
  annoy.metric = "cosine",
  verbose = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay_name}{Assay name. Can provide two assay names to perform joint clustering across both}

\item{do.scale}{scale}

\item{do.center}{center}

\item{algorithm}{clustering algorithm}

\item{resolution}{clustering resolution}

\item{n.neighbors}{neighbors for umap}

\item{npcs.pca}{Total Number of PCs to compute and store (50 by default)}

\item{features.pca}{One of 'all', 'variable', or a vector of features to include in dimensionality reduction. Defaults to 'all'.}

\item{nvar.features}{Number of variable features if \code{features.pca='variable'}}

\item{dims}{Number of reduced dimensions to use for FindNeighbors and UMAP}

\item{k.param}{Defines k for the k-nearest neighbor algorithm}

\item{suffix}{Suffix name to add to the PCA, UMAP, and clusters}

\item{PCA_name}{Name to store PCA dimred}

\item{UMAP_name}{Name to store UMAP dimred}

\item{cluster_name}{Name to store seurat clusters}

\item{umap.metric}{Metric for \link[Seurat:RunUMAP]{Seurat::RunUMAP}}

\item{annoy.metric}{Metric for \link[Seurat:FindNeighbors]{Seurat::FindNeighbors}}

\item{verbose}{Message verbosity}
}
\value{
SingleCellExperiment obj
}
\description{
Cluster using Seurat
}
