% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cnaHeatmap}
\alias{cnaHeatmap}
\title{Plot copy number heatmap}
\usage{
cnaHeatmap(
  sce,
  assay_name = "state",
  clone_name = NULL,
  cell_order = NULL,
  cluster_rows = NULL,
  log2 = FALSE,
  center = FALSE,
  scale = c("none", "cells", "bins", "both"),
  clustering_results = NULL,
  label_genes = NULL,
  col_fun = NULL,
  col_clones = NULL,
  cluster_clones = FALSE,
  legend_name = assay_name,
  clust_annot = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{assay_name}{Name of the assay to plot}

\item{clone_name}{Name of clone_id column in sce object}

\item{cell_order}{Optional: Order of the cells}

\item{cluster_rows}{If the value is a logical, it controls whether to make cluster on rows. The value can also be a \code{stats::hclust} or a \code{stats::dendrogram} which already contains clustering. Check https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#clustering.}

\item{log2}{Logical: Log2 transform the matrix prior to plotting}

\item{center}{logical: center the matrix prior to plotting}

\item{scale}{One of \verb{'cells', 'bins', 'both' or 'none'}. Determines what kind of scaling is done.}

\item{clustering_results}{Clustering results to provide to inform cell ordering and cluster labelling. From \link{perform_umap_clustering}}

\item{label_genes}{Optional: Vector of gene names to label in the heatmap. Note: \link{overlap_genes} must be run prior to labelling genes.}

\item{col_fun}{Color mapping function from \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}}

\item{col_clones}{Optional: A named vector (or unnamed) of clone colors.}

\item{cluster_clones}{Logical: Whether or not to order the clones by clustering}

\item{legend_name}{Name of the legend}

\item{clust_annot}{Annotate cluster and sample labels}

\item{verbose}{Logical: Message verbosity}

\item{...}{Additional parameters that can be passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}}
}
\value{
A heatmap
}
\description{
Plot copy number heatmap
}
