% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cnaHeatmap}
\alias{cnaHeatmap}
\title{Plot copy number heatmap}
\usage{
cnaHeatmap(
  sce,
  assay_name = "state",
  cell_order = NULL,
  log2 = FALSE,
  scale = FALSE,
  clustering_results = NULL,
  col_fun = NULL,
  legend_name = assay_name,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{assay_name}{Name of the assay to plot}

\item{cell_order}{Optional: Order of the cells}

\item{log2}{Logical: Log2 transform the matrix prior to plotting}

\item{scale}{Logical: Scale and center the matrix before plotting}

\item{clustering_results}{Clustering results to provide to inform cell ordering and cluster labelling. From \link{perform_umap_clustering}}

\item{col_fun}{Color mapping function from \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}}

\item{legend_name}{Name of the legend}

\item{...}{Additional parameters that can be passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}}
}
\value{
A heatmap
}
\description{
Plot copy number heatmap
}
