% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{do_qc}
\alias{do_qc}
\title{Basic QC}
\usage{
do_qc(sce, assay_name = "counts", plot = TRUE)
}
\arguments{
\item{sce}{sce}

\item{assay_name}{assay}

\item{plot}{logical plot}
}
\value{
sce
}
\description{
Basic QC
}
