% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_snps.R
\name{EM2}
\alias{EM2}
\title{Iterative phasing and theta_hat estimation}
\usage{
EM2(ref_table, alt_table, max_iter = 50, seed = 3, phases = NULL, tol = 0.001)
}
\arguments{
\item{ref_table}{A SNP by cell read count matrix/ spare matrix for the reference alleles.}

\item{alt_table}{A SNP by cell read count matrix/ spare matrix for the alternative alleles.}

\item{max_iter}{An integer of maximum iteration number.}

\item{seed}{An integer of random seed number for EM initialization.}

\item{phases}{SNP phasing to use as priors}

\item{tol}{maximum liklihood stopping tolerance}
}
\value{
A list of estimated indicators (I_hat) for each SNP and estimated major haplotype proportion (theta_hat) for each cell in one region. I_hat is the phasing result indicating whether reference allele is on the major haplotype for each SNP. Theta_hat represents the CNV states for each cell. A cell is considered as a CNV carrier if its theta_hat depart from 0.5.
}
\description{
Iterative phasing and theta_hat estimation
}
