% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{get_cytobands}
\alias{get_cytobands}
\title{Get genome cytobands}
\usage{
get_cytobands(genome = "hg38")
}
\arguments{
\item{genome}{Genome version (hg38 or hg19)}
}
\value{
Dataframe of genome cytobands
}
\description{
Get genome cytobands
}
\examples{
hg38_cyto <- get_cytobands('hg38')
}
