% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{leiden_wrapper}
\alias{leiden_wrapper}
\title{Wrapper for the Leiden Algorithm}
\usage{
leiden_wrapper(adj_mat, resolution = 1)
}
\arguments{
\item{adj_mat}{Adjacency matrix}

\item{resolution}{Resolution paramter}
}
\value{
cluster memberships
}
\description{
Wrapper for the Leiden Algorithm
}
