% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{load_atac_bins}
\alias{load_atac_bins}
\title{Load atac binned depth data}
\usage{
load_atac_bins(
  directory,
  ArchR_Proj = NULL,
  bins = NULL,
  BPPARAM = BiocParallel::bpparam(),
  ncores = 1,
  save_to = NULL,
  verbose = FALSE,
  save_as = c("sce", "adata", "seurat")
)
}
\arguments{
\item{directory}{Folder containing samples}

\item{ArchR_Proj}{Optional: ArchR project with matching cells and metadata}

\item{bins}{Optional: GRanges bins object}

\item{BPPARAM}{Options to pass to \code{bplapply} for data loading. Provides minor speedup if loading many samples}

\item{ncores}{number of cores for parallel evaluation (requires \link{pbmcapply} package)}

\item{save_to}{File path in which to save the final output. Note: Will still return the sce object for downstream analysis.}

\item{verbose}{Message verbosity (TRUE/FALSE)}

\item{save_as}{NOT IMPLEMENTED YET. Select file formats to save the object. Can provide multiple values}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Loads binned atac reads, merges cell-wise and bin-wise metadata, and performs QC.
}
