% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_atac_bins.R
\name{load_atac_bins}
\alias{load_atac_bins}
\title{Load atac binned depth data}
\usage{
load_atac_bins(bin_dir, sample_id, bins = NULL, save_to = NULL, verbose = TRUE)
}
\arguments{
\item{bin_dir}{Directory with the bin counts}

\item{sample_id}{Sample ID}

\item{bins}{Optional: GRanges bins object}

\item{save_to}{File path in which to save the final output. Note: Will still return the sce object for downstream analysis.}

\item{verbose}{Message verbosity}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Loads binned atac reads, merges cell-wise and bin-wise metadata, and performs QC.
}
