% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_correction.R
\name{perform_gc_cor}
\alias{perform_gc_cor}
\title{Perform GC Correction}
\usage{
perform_gc_cor(
  mat,
  gc,
  valid_mat = NULL,
  method = c("modal", "copykit", "loess"),
  bpparam,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mat}{Count matrix for GC correction}

\item{gc}{GC corresponding to bins (rows) in the matrix}

\item{valid_mat}{Matrix of TRUE/FALSE for valid bins. If none provided defaults to all TRUE}

\item{method}{Specifies the type of GC correction to perform. One of \verb{'modal', 'copykit', or 'loess'}}

\item{bpparam}{BiocParallel params}

\item{verbose}{Message verbosity (TRUE/FALSE)}

\item{...}{Additional arguments to be passed to GC correction methods}
}
\value{
Sparse matrix of corrected counts
}
\description{
Performs GC correction using over a matrix of cell counts
}
\details{
Note: If using \code{modal} must pass \code{results="counts"} as an argument. Uses \code{\link[pbmcapply:pbmclapply]{pbmcapply::pbmclapply()}} for parallelization.
}
