% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cell_psuedobulk_cna}
\alias{plot_cell_psuedobulk_cna}
\title{Plot Psuedobulk cell CNA profiles}
\usage{
plot_cell_psuedobulk_cna(
  sce,
  assay_name,
  group_var = "all",
  FUN = mean,
  col_fun = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment Object}

\item{assay_name}{Name of the assay}

\item{group_var}{Grouping variable to pseudobulk across. Default: "all"}

\item{col_fun}{Color mapping function from \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}}
}
\value{
ggplot
}
\description{
Plot Psuedobulk cell CNA profiles
}
