% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vartrix.R
\name{read_vartrix}
\alias{read_vartrix}
\title{Read Vatrix Data}
\usage{
read_vartrix(
  dir_path = NULL,
  mtx_ref = NULL,
  mtx_alt = NULL,
  barcodes = NULL,
  variants = NULL,
  input_vcf = NULL,
  phased_vcf = NULL,
  verbose = FALSE,
  keep_barcodes = NULL,
  blacklist = NULL,
  min_counts = 1
)
}
\arguments{
\item{dir_path}{Path to \code{Vartrix} outputs directory}

\item{mtx_ref}{Path to reference allele \code{.mtx} file}

\item{mtx_alt}{Path to alternate allele \code{.mtx} file}

\item{barcodes}{Path to barcodes file}

\item{variants}{Path to variants file}

\item{input_vcf}{Germline VCF before phasing}

\item{phased_vcf}{Phased VCF}

\item{verbose}{Verbosity}

\item{keep_barcodes}{Vector of cell barcodes to keep. Will perform an intersect with this list.}

\item{blacklist}{Optional blacklist to remove SNPs}

\item{min_counts}{Minimum count number across all cells to keep a SNP}
}
\value{
A \code{SingleCellExperiment} object
}
\description{
Reads in \code{Vartrix} scSNP pileups. Assumes Vartrix is run in \verb{-s coverage} mode, outputting barcodes and variants table. If no \code{dir_path} is provided, then user must provide individual file paths for each file. If phased VCF and input VCF is provided this function will also remove any non-heterozygous loci, and indels.
}
