% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scatools.R
\name{run_scatools}
\alias{run_scatools}
\title{Run scATAC Copy Number Analysis}
\usage{
run_scatools(
  sample_id,
  fragment_file,
  cells = NULL,
  bins = bins_10mb,
  bin_name = prettyMb(getmode(width(bins))),
  blacklist = NULL,
  outdir = sample_id,
  rmsb_size = 0.1 * getmode(width(bins)),
  gc_range = c(0.3, 0.8),
  overwrite = FALSE,
  verbose = TRUE,
  save_h5ad = TRUE,
  ncores = 1,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{sample_id}{Sample id}

\item{fragment_file}{Fragments file}

\item{cells}{A vector of cell barcodes. If not provided will use all barcodes}

\item{bins}{Genomic ranges object with bins to use}

\item{bin_name}{bin_name Name of the bins (e.g. \code{'10Mb'}, \code{'500Kb'}, \code{'chr_arm'}). If not provided is automatically detected based on binwidth.}

\item{blacklist}{Genome blacklist regions to filter against}

\item{outdir}{Output directory}

\item{rmsb_size}{Remove small bins below a given size. Useful in cases where small bins are leftover at the ends of chromosomes. Defaults to 10\% of the binwidth.}

\item{gc_range}{GC range for bins to keep. Removes large GC outliers}

\item{overwrite}{Whether to overwrite previous results}

\item{verbose}{Verbosity}

\item{save_h5ad}{Logical. Whether to save raw and processed h5ad files. Requires packages \code{zellkonverter} and \code{anndata}}

\item{ncores}{Number of cores}

\item{BPPARAM}{BiocParallel Parameters}
}
\value{
\code{SingleCellExperiment} object
}
\description{
Convenience wrapper to run a sample
}
