% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scatools.R
\name{run_scatools}
\alias{run_scatools}
\title{Run scATAC Copy Number Analysis}
\usage{
run_scatools(
  sample_id,
  fragment_file,
  cells = NULL,
  bins = bins_10mb,
  bin_name = prettyMb(getmode(width(bins))),
  blacklist = NULL,
  outdir = sample_id,
  rmsb_size = 0.1 * getmode(width(bins)),
  gc_range = c(0.3, 0.8),
  overwrite = FALSE,
  verbose = TRUE,
  save_h5ad = TRUE,
  min_bin_counts = 1,
  min_bin_prop = 0.95,
  min_cell_counts = min_bin_counts,
  min_cell_prop = min_bin_prop,
  ncores = 1,
  bpparam = BiocParallel::SerialParam()
)
}
\arguments{
\item{sample_id}{Sample id}

\item{fragment_file}{Fragments file}

\item{cells}{A vector of cell barcodes. If not provided will use all barcodes}

\item{bins}{Genomic ranges object with bins to use}

\item{bin_name}{bin_name Name of the bins (e.g. \code{'10Mb'}, \code{'500Kb'}, \code{'chr_arm'}). If not provided is automatically detected based on binwidth.}

\item{blacklist}{Genome blacklist regions to filter against}

\item{outdir}{Output directory}

\item{rmsb_size}{Remove small bins below a given size. Useful in cases where small bins are leftover at the ends of chromosomes. Defaults to 10\% of the binwidth.}

\item{gc_range}{GC range for bins to keep. Removes large GC outliers}

\item{overwrite}{Whether to overwrite previous binned counts}

\item{verbose}{Verbosity}

\item{save_h5ad}{Logical. Whether to save raw and processed h5ad files. Requires packages \code{zellkonverter} and \code{anndata}}

\item{min_bin_counts}{A bin requires at least \code{min_bin_counts} across \code{min_bin_prop} proportion of cells to be kept}

\item{min_bin_prop}{Minimum proportion of cells with at least \code{min_bin_counts} per bin in order to keep a bin}

\item{min_cell_counts}{A cell requires at least \code{min_cell_counts} across \code{min_cell_prop} proportion of bins to be kept}

\item{min_cell_prop}{Minimum proportion of bins with at least \code{min_cell_counts} per cell in order to keep a cell}

\item{ncores}{Number of cores}

\item{bpparam}{BiocParallel Parameters}
}
\value{
\code{SingleCellExperiment} object
}
\description{
Convenience wrapper to run a sample
}
